+++ packages/packsquash/src/pack_file/shader_file/example_false_positive_parse_error.glsl
@@ -1,0 +1,8 @@
+// unresolvable.glsl may be resolvable by the game, but
+// not by us, and contain preprocessor directives that
+// may render otherwise invalid GLSL source valid. In
+// the example below, unresolvable.glsl may define
+// SEMICOLON to be ;
+#moj_import <unresolvable.glsl>
+
+vec2 var SEMICOLON