+++ packages/packsquash/src/pack_file/shader_file/example_non_transformable.glsl
@@ -2,7 +2,8 @@
 
 void main() {
     // Preprocessor directives outside of external declaration position
-    // (i.e., within statements or other GLSL symbols) cause PackSquash
-    // to fall back to no source transformation
-	#moj_import <mist.glsl>
+    // (i.e., within statements or other GLSL symbols) or unresolvable
+    // #moj_import cause PackSquash to fall back to no source
+    // transformation
+    #moj_import <mist.glsl>
 }