+++ packages/packsquash/src/pack_file/shader_file/tests.rs
@@ -8,15 +8,18 @@ use super::*;
 
 static FRAGMENT_SHADER_DATA: &[u8] = include_bytes!("example.fsh");
 static NON_TRANSFORMABLE_SHADER_DATA: &[u8] = include_bytes!("example_non_transformable.glsl");
+static FALSE_POSITIVE_PARSE_ERROR: &[u8] = include_bytes!("example_false_positive_parse_error.glsl");
 
 /// Processes the given input data as a [ShaderFile], using the provided settings,
 /// expecting a successful result that equals the expected string.
+#[allow(clippy::too_many_arguments)]
 async fn successful_process_test<T: Extractable<TranslationUnit> + PartialEq + Debug + 'static>(
 	input_data: &[u8],
 	add_bom: bool,
 	settings: ShaderFileOptions,
 	is_vertex_or_fragment_shader: bool,
 	is_top_level_translation_unit: bool,
+	expect_parseable_input_and_equal_ast_after_processing: bool,
 	expect_smaller_file_size: bool,
 	expect_output_equal_to_input: bool
 ) {
@@ -24,11 +27,7 @@ async fn successful_process_test<T: Extractable<TranslationUnit> + PartialEq + D
 
 	let input_data_str = std::str::from_utf8(input_data);
 
-	let input_ast = shader_parser
-		.parse::<T>(input_data, is_top_level_translation_unit)
-		.expect("The test input data should be a valid GLSL symbol");
-
-	let input_data = if add_bom {
+	let input_data_with_maybe_bom = if add_bom {
 		let mut input_data = input_data.to_vec();
 		input_data.splice(0..0, BOM_UTF8);
 		Cow::Owned(input_data)
@@ -37,8 +36,8 @@ async fn successful_process_test<T: Extractable<TranslationUnit> + PartialEq + D
 	};
 
 	let data_stream = ShaderFile {
-		read: Builder::new().read(&input_data).build(),
-		file_length_hint: input_data.len(),
+		read: Builder::new().read(&input_data_with_maybe_bom).build(),
+		file_length_hint: input_data_with_maybe_bom.len(),
 		is_vertex_or_fragment_shader,
 		optimization_settings: settings
 	}
@@ -49,7 +48,7 @@ async fn successful_process_test<T: Extractable<TranslationUnit> + PartialEq + D
 		.collect()
 		.await;
 
-	let mut data = Vec::with_capacity(input_data.len());
+	let mut data = Vec::with_capacity(input_data_with_maybe_bom.len());
 	for (_, partial_data) in process_result {
 		data.extend_from_slice(&partial_data);
 	}
@@ -63,14 +62,20 @@ async fn successful_process_test<T: Extractable<TranslationUnit> + PartialEq + D
 
 	assert!(!data.is_empty(), "Some data was expected for this input");
 
-	let processed_ast = shader_parser
-		.parse::<T>(&data, is_top_level_translation_unit)
-		.expect("The result should be a valid GLSL symbol");
+	if expect_parseable_input_and_equal_ast_after_processing {
+		let input_ast = shader_parser
+			.parse::<T>(input_data, is_top_level_translation_unit)
+			.expect("The test input data should be a valid GLSL symbol");
 
-	assert_eq!(
-		input_ast, processed_ast,
-		"The processed translation unit should represent the same AST as the input"
-	);
+		let processed_ast = shader_parser
+			.parse::<T>(&data, is_top_level_translation_unit)
+			.expect("The result should be a valid GLSL symbol");
+
+		assert_eq!(
+			input_ast, processed_ast,
+			"The processed translation unit should represent the same AST as the input"
+		);
+	}
 
 	assert!(
 		!expect_smaller_file_size || data.len() < input_data.len(),
@@ -94,6 +99,7 @@ async fn minifying_works() {
 		},
 		true,  // Fragment shader
 		true,  // Top-level TU
+		true,  // The output AST should match the input AST
 		true,  // Smaller size
 		false  // Different output text
 	)
@@ -111,6 +117,7 @@ async fn minifying_with_bom_works() {
 		},
 		true,  // Fragment shader
 		true,  // Top-level TU
+		true,  // The output AST should match the input AST
 		true,  // Smaller size
 		false  // Different output text
 	)
@@ -128,6 +135,7 @@ async fn prettifying_works() {
 		},
 		true,  // Fragment shader
 		true,  // Top-level TU
+		true,  // The output AST should match the input AST
 		false, // Same/bigger size
 		false  // Different output text
 	)
@@ -145,6 +153,7 @@ async fn passthrough_works() {
 		},
 		true,  // Fragment shader
 		true,  // Top-level TU
+		true,  // The output AST should match the input AST
 		false, // Same size
 		true   // Same output text
 	)
@@ -169,7 +178,7 @@ async fn invalid_input_is_handled() {
 }
 
 #[tokio::test]
-async fn minifying_is_averted_when_preprocessor_directives_would_be_lost() {
+async fn minifying_is_averted_when_preprocessor_expansion_would_be_incomplete() {
 	successful_process_test::<TranslationUnit>(
 		NON_TRANSFORMABLE_SHADER_DATA,
 		false, // No BOM
@@ -179,6 +188,25 @@ async fn minifying_is_averted_when_preprocessor_directives_would_be_lost() {
 		},
 		true,  // Fragment shader
 		true,  // Top-level TU
+		true,  // The output AST should match the input AST
+		false, // Same size
+		true   // Same output text
+	)
+	.await
+}
+
+#[tokio::test]
+async fn tentative_parsing_errors_are_not_fatal() {
+	successful_process_test::<TranslationUnit>(
+		FALSE_POSITIVE_PARSE_ERROR,
+		false, // No BOM
+		ShaderFileOptions {
+			source_transformation_strategy: ShaderSourceTransformationStrategy::Minify,
+			..Default::default()
+		},
+		true,  // Fragment shader
+		true,  // Top-level TU
+		false, // The input AST cannot be constructed
 		false, // Same size
 		true   // Same output text
 	)
@@ -196,6 +224,7 @@ async fn include_shaders_may_not_be_translation_units() {
 		},
 		false, // Include shader
 		false, // Not top-level
+		true,  // The output AST should match the input AST
 		true,  // Smaller size
 		false  // Different output text
 	)
@@ -213,6 +242,7 @@ async fn include_shaders_with_preprocessor_directives_are_passed_through() {
 		},
 		false, // Include shader
 		false, // Not top-level
+		true,  // The output AST should match the input AST
 		false, // Smaller size
 		true   // Same output text
 	)